+++ src/Screens/TemplateScreen/__TC__/ScreenButtons.tsx
@@ -1,7 +1,6 @@
 import React, { useMemo } from 'react';
 import { useLocalSearchParams } from 'expo-router';
 
-import { navigate } from '@Globals/NavigationControler';
 import ConfigService from '@Services/ConfigService';
 import ProjectService from '@Services/ProjectService';
 import CacheService from '@Services/CacheService';
@@ -35,19 +34,6 @@ export default function ScreenButtons(props: {
       buttons={[
         <Button.RoundedIcon
           key="1"
-          iconName="arrow-back"
-          showPlusSign={false}
-          buttonDiameter={60}
-          onPress={() => navigate('HOME SCOPE')}
-          theme={{
-            font: theme.font,
-            font_Pressed: theme.backgroud,
-            background: theme.backgroud,
-            background_Pressed: theme.background_active,
-          }}
-        />,
-        <Button.RoundedIcon
-          key="2"
           iconName="square"
           showPlusSign={true}
           buttonDiameter={60}